Slight improvement to html-classes.
This change causes the html-classes extra to perform class lookups for
"img start" "img end" and "img start end" as well the usual "img" from
before. "start" indicates the image was at the start of a line in the
markdown. "end" indicates the end of a line.

I like to use the position of an image to influence its output. For
instance, an image at the beginning of a line of text might float left
of the following text and an image on a line by itself might be framed
differently.