"nofollow" "target-blank-lines" respect fragments
Both the "nofollow" and "target-blank-lines" extras need to ignore links
that point to fragments within the same document. For "nofollow", this
is because adding a rel="nofollow" attribute would be unnecessary, and
for "target-blank-lines", because this would cause links like the ones
generated by the "footnotes" extra to open unexpectedly in new tabs.