Don't auto link patterns surrounded by """.
This allows the prevention of `[WikiWikiWeb](http://wiki.c2.com/?WikiWikiWeb)`
being transformed to `<a href="http://wiki.c2.com/?&lt;a href=&quot;aw.cgi?page=WikiWikiWeb&quot;&gt;WikiWikiWeb&lt;/a&gt;">WikiWikiWeb</a>` when
the pattern linker sees the second WikiCase word by surrounding it with triple quotes: so `[WikiWikiWeb](http://wiki.c2.com/?"""WikiWikiWeb""")`
is now transformed to `<a href="http://wiki.c2.com/?WikiWikiWeb">WikiWikiWeb</a>`.

The function responsible, _do_link_patterns(), already checked for the cases that the pattern was inside square brackets, or
was preceeded by `](` or suffixed by `")`, and skips expanding in these cases. This addition simply adds a check that the pattern is not surrounded
by triple quotes; the triple quotes are stripped from the output.