Fix code surrounded by blank lines inside quoted fenced code blocks being treated as another code block.
If you were using the `fenced-code-blocks` extension and you had code that was indented (eg: inside an if statement) and
surrounded by blank lines, it would be detected as a code block, regardless of whether it was already inside a code block.

EG:
> ```python
> if True:
> 	print()
>
> 	print()  # this line would register as a code block
>
> 	print()
> ```

This is because the regex look-ahead being used to check for a `</code>` tag would only check the next HTML tag it found,
which would usually be a `<span>` tag if syntax highlighting was being used.

The new look-ahead checks the next HTML tag that isn't a `<span>` tag to see if it's a `</code>` tag