Fix #426
The problem was that fenced code blocks would not be detected if placed inside a list (ul/ol) due to
the leading indentation before the opening code fence.
This commit tweaks the `_fenced_code_block_re` to ignore leading space/tab indents before an opening fence.

`_code_block_sub` has also been altered to maintain the leading indentation when inserting code hashes as
to not break the list (ul/ol) that the fenced code block is in.