Fix #489.
The issue was fenced code blocks where the first line of code was indented from the fence
indentation level. The codeblock would be uniformly outdented past the level that the
fence was at, overriding the `lead_indent` variable set for the fence.

Then, when the codeblock was re-indented, it would be uniformly re-indented to `lead_indent`,
meaning that the fence would gain aditional indentation, making it look like a classic code block