Custom extras initial proof of concept.
The way this works is by creating subclasses of the `Extra` class. These subclasses will have an order and a name. The name is the same one
specified in the `extras` list/dict given to the Markdown init function.

The order will be at which point the function will be executed. This is done by attaching the extra to a "Stage", a distinct step in the
markdown process (eg: forming paragraphs, processing links... etc). You can set the extra to run before or after the stage.

At the moment, extras are automatically registered, activated and executed by the Markdown class.

TODO:
* More elegant way to register and init extras
* Optimise `Stage.mark`
* Convert more extras to new class format